refactor(frontend): rename components as jsx
